Initial Fabric Server Test Workflow
This is an initial attempt at setting up a fabric test server workflow. Started with fabric as it seemes much less of a pain to do. This matrix only runs if the user has a github secret called MC_EULA_ACCEPTED set to true. This is just to avoid any legal bother in the unlikely event that happens.